# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: mailtrap-local API
  version: 0.1.0
  summary: Local email sandbox + catcher for individual developers — JSON `/api/v1/*`.
  description: |
    mailtrap-local is a local email sandbox: it runs on a developer's
    machine, catches every message their app sends, and exposes the
    sandbox over a JSON API at `/api/v1/*`. Field names use snake_case
    on the wire (`from`, `to`, `message_id`, `tags`) — the same
    convention used across the rest of the Mailtrap toolchain (sandbox
    API, `mailtrap-cli`).

    Notable features:

    - **Single-string categories.** mailtrap-local stores one `category`
      string per message (read at ingest from `X-MT-Category` — Mailtrap's
      SMTP header convention — with `Category` as a fallback). The wire
      format exposes this as `tags: string[]`, but the array is always
      0 or 1 element (the category).
    - **Mailtrap cloud + SMTP-relay extensions.** Endpoints under
      `/cloud_connection` and `/relay_connection` configure forwarding to
      a Mailtrap cloud sandbox (`POST /message/:id/send_to_cloud`) or a
      generic SMTP relay (`POST /message/:id/release`).

    Default bind is loopback only — `127.0.0.1:3550` for the API. There is no
    auth: this is a single-user local tool.

  contact:
    name: Mailtrap
    url: https://mailtrap.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:3550
...
paths:
...
  /api/v1/messages:
    get:
      tags: [messages]
      summary: List messages
      description: |
        Returns messages from the sandbox ordered newest-first. The `tags` field in
        the envelope  lists distinct **categories** across the
        *unfiltered* sandbox — useful for rendering a category picker without a
        second roundtrip.
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/Start'
        - $ref: '#/components/parameters/Limit'
        - name: category
          in: query
          description: |
            **mailtrap-local extension.** Narrows the result set to messages whose
            stored category equals this value (exact match). The envelope's `tags`
            still reports the unfiltered distinct list so a picker stays populated.
          required: false
          schema:
            type: string
          example: welcome
      responses:
        '200':
          description: Paginated list of message summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'

    put:
      tags: [messages]
      summary: Mark messages as read or unread
      description: |
        Toggles read state in bulk. With `ids` set, marks just those messages;
        without `ids`, marks **all** messages.
      operationId: updateMessagesRead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [read]
              properties:
                read:
                  type: boolean
                  description: True = mark as read, false = mark as unread.
                ids:
                  type: array
                  description: Optional. When omitted, applies to all messages.
                  items:
                    type: string
              example:
                read: true
                ids: ["abc123", "def456"]
      responses:
        '200':
          description: '"ok" plain-text body.'
          content:
            text/plain:
              schema:
                type: string
                example: ok

    delete:
      tags: [messages]
      summary: Delete messages
      description: |
        With a non-empty `ids` array in the body, deletes those specific messages.
        Without a body (or with empty `ids`), deletes **all** messages in the sandbox.
      operationId: deleteMessages
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
              example:
                ids: ["abc123"]
      responses:
        '200':
          description: '"ok" plain-text body.'
          content:
            text/plain:
              schema:
                type: string
                example: ok

  # ---- Search ----------------------------------------------------------
...
  /api/v1/message/{id}:
    parameters:
      - $ref: '#/components/parameters/MessageID'
    get:
      tags: [messages]
      summary: Get message detail
      description: |
        Returns the full parsed message including bodies and attachment metadata.
        **Side effect:** marks the message as read (matches the documented behavior).

        Pass the literal `latest` as `id` to fetch the most recent message
        (convention).
      operationId: getMessage
      responses:
        '200':
          description: Full message detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/message/{id}/html_check:
    parameters:
      - $ref: '#/components/parameters/MessageID'
    get:
      tags: [messages]
      summary: HTML email-client compatibility check
      description: |
        Runs a rule engine over the message HTML and returns issues
        with caniemail.com client/version data plus market-share
        weighted support stats. Cached per-message (HTML is immutable).

        Status values:
          - `success`            — report ready (see schema)
          - `no_html`            — message has no HTML body
          - `size_limit_exceeded` — HTML exceeds the size cap
          - `error`              — validator hit an unexpected error
      operationId: htmlCheck
      responses:
        '200':
          description: Report (or status placeholder).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true

  # ---- Internal ingest -------------------------------------------------
...
components:
  parameters:
    Start:
      name: start
      in: query
      description: Pagination offset. Clamped to `[0, 1_000_000]`.
      required: false
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: Page size. Clamped to `[1, 200]`. Default 50.
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 200
        default: 50
    MessageID:
      name: id
      in: path
      required: true
      description: |
        Message database ID, or the literal string `latest` for the newest message.
      schema:
        type: string
  responses:
    NotFound:
      description: Message not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Address:
      type: object
      description: Email address with optional display name.
      properties:
        name:
          type: string
          description: Display name (may be empty).
        address:
          type: string
          format: email
      example:
        name: Sender
        address: sender@example.com
    AttachmentSummary:
      type: object
      description: Attachment metadata. Bytes are fetched via `/message/{id}/part/{part_id}`. Matches `Attachment` (response shape).
      properties:
        part_id:
          type: string
        file_name:
          type: string
        content_type:
          type: string
        content_id:
          type: string
        size:
          type: integer
          format: uint64
        checksums:
          type: object
          properties:
            md5: { type: string }
            sha1: { type: string }
            sha256: { type: string }
    ListUnsubscribe:
      type: object
      description: |
        Parsed `List-Unsubscribe` / `List-Unsubscribe-Post` headers, when present.
        Empty stub when the message has none.
      properties:
        header:
          type: string
        links:
          type: array
          items:
            type: string
        errors:
          type: string
        header_post:
          type: string
    MessageSummary:
      type: object
      description: |
        List-row shape returned by `GET /messages`.

        **Note on `tags`:** the array is always 0 or 1 element (the message's
        category). mailtrap-local stores a single category per message — see
        the API description for rationale.
      properties:
        id:
          type: string
          description: Database ID.
        message_id:
          type: string
          description: RFC 5322 Message-ID header.
        read:
          type: boolean
        from:
          $ref: '#/components/schemas/Address'
        to:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        cc:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        bcc:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        reply_to:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        subject:
          type: string
        created:
          type: string
          format: date-time
          description: Received RFC3339Nano date & time.
        username:
          type: string
          description: Always empty in mailtrap-local — there is no SMTP AUTH (single-user local tool).
        tags:
          type: array
          maxItems: 1
          items:
            type: string
          description: 0- or 1-element array containing the message's category .
        size:
          type: integer
          format: uint64
        attachments:
          type: integer
          format: uint64
          description: Count of non-inline attachments.
        snippet:
          type: string
          description: Up to 250 characters of the text body.
    Message:
      type: object
      description: |
        Full message detail returned by `GET /message/{id}`. Includes two
        mailtrap-local extensions documented at the bottom (`envelope_from`
        and `envelope_to`); clients should ignore unknown keys.
      properties:
        id:
          type: string
        message_id:
          type: string
        from:
          $ref: '#/components/schemas/Address'
        to:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        cc:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        bcc:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        reply_to:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        return_path:
          type: string
        subject:
          type: string
        list_unsubscribe:
          $ref: '#/components/schemas/ListUnsubscribe'
        date:
          type: string
          format: date-time
          description: Header `Date` if parseable, otherwise ingest time.
        tags:
          type: array
          maxItems: 1
          items:
            type: string
          description: 0- or 1-element array containing the message's category .
        username:
          type: string
        text:
          type: string
        html:
          type: string
        size:
          type: integer
          format: uint64
        inline:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentSummary'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentSummary'
        envelope_from:
          type: string
          description: '**mailtrap-local extension.** SMTP `MAIL FROM` envelope sender. Surfaced in the Tech Info tab.'
        envelope_to:
          type: array
          items:
            type: string
          description: '**mailtrap-local extension.** SMTP `RCPT TO` envelope recipients.'
    MessagesResponse:
      type: object
      description: Envelope returned by `GET /messages`.
      properties:
        total:
          type: integer
          format: int64
          description: Total messages matching the current filter (e.g. `?category=`).
        unread:
          type: integer
          format: int64
        count:
          type: integer
          format: int64
          description: Number of summaries in this page.
        messages_count:
          type: integer
          format: int64
          description: Mirror of `total` .
        messages_unread:
          type: integer
          format: int64
          description: Mirror of `unread` .
        start:
          type: integer
          format: int64
        tags:
          type: array
          items:
            type: string
          description: |
            Distinct **categories** across the *unfiltered* sandbox (named
            `tags` in the response for  wire-format compatibility).
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageSummary'
...
    Error:
      type: object
      description: Plain JSON error envelope used across the API.
      properties:
        error:
          type: string
      example:
        error: Message not found